+++ src/flyingcircus/service/amplify.py
@@ -1,0 +1,12 @@
+"""General-use classes to interact with the Amplify service through CloudFormation.
+
+See Also:
+    `AWS developer guide for Amplify
+    <https://docs.aws.amazon.com/amplify/latest/userguide/getting-started.html>`_
+"""
+
+# noinspection PyUnresolvedReferences
+from .._raw import amplify as _raw
+
+# noinspection PyUnresolvedReferences
+from .._raw.amplify import *