+++ src/flyingcircus/service/backup.py
@@ -1,0 +1,12 @@
+"""General-use classes to interact with the Backup service through CloudFormation.
+
+See Also:
+    `AWS developer guide for Backup
+    <https://docs.aws.amazon.com/aws-backup/latest/devguide/index.html>`_
+"""
+
+# noinspection PyUnresolvedReferences
+from .._raw import backup as _raw
+
+# noinspection PyUnresolvedReferences
+from .._raw.backup import *