+++ src/flyingcircus/service/iotthingsgraph.py
@@ -1,0 +1,12 @@
+"""General-use classes to interact with the IoTThingsGraph service through CloudFormation.
+
+See Also:
+    `AWS developer guide for IoTThingsGraph
+    <IoTThingsGraph>`_
+"""
+
+# noinspection PyUnresolvedReferences
+from .._raw import iotthingsgraph as _raw
+
+# noinspection PyUnresolvedReferences
+from .._raw.iotthingsgraph import *