+++ src/flyingcircus/service/medialive.py
@@ -1,0 +1,12 @@
+"""General-use classes to interact with the MediaLive service through CloudFormation.
+
+See Also:
+    `AWS developer guide for MediaLive
+    <https://docs.aws.amazon.com/pinpoint/latest/userguide/welcome.html>`_
+"""
+
+# noinspection PyUnresolvedReferences
+from .._raw import medialive as _raw
+
+# noinspection PyUnresolvedReferences
+from .._raw.medialive import *