+++ src/flyingcircus/service/msk.py
@@ -1,0 +1,12 @@
+"""General-use classes to interact with the MSK service through CloudFormation.
+
+See Also:
+    `AWS developer guide for MSK
+    <https://docs.aws.amazon.com/pinpoint/latest/userguide/welcome.html>`_
+"""
+
+# noinspection PyUnresolvedReferences
+from .._raw import msk as _raw
+
+# noinspection PyUnresolvedReferences
+from .._raw.msk import *