+++ src/flyingcircus/service/pinpoint.py
@@ -1,0 +1,12 @@
+"""General-use classes to interact with the Pinpoint service through CloudFormation.
+
+See Also:
+    `AWS developer guide for Pinpoint
+    <https://docs.aws.amazon.com/pinpoint/latest/userguide/welcome.html>`_
+"""
+
+# noinspection PyUnresolvedReferences
+from .._raw import pinpoint as _raw
+
+# noinspection PyUnresolvedReferences
+from .._raw.pinpoint import *