+++ src/flyingcircus/service/securityhub.py
@@ -1,0 +1,12 @@
+"""General-use classes to interact with the SecurityHub service through CloudFormation.
+
+See Also:
+    `AWS developer guide for SecurityHub
+    <https://docs.aws.amazon.com/securityhub/latest/userguide/index.html>`_
+"""
+
+# noinspection PyUnresolvedReferences
+from .._raw import securityhub as _raw
+
+# noinspection PyUnresolvedReferences
+from .._raw.securityhub import *